+++ src/actions/cardData.js
@@ -1,0 +1,1 @@
+export const SET_DATA_CARD = 'SET_DATA_CARD';