+++ src/actions/favorites.js
@@ -1,1 +1,0 @@
-export const SET_TO_FAVORITE = 'SET_TO_FAVORITE';