+++ src/screens/app/Detail/index.js
@@ -4,24 +4,40 @@ import { View, Text, StatusBar, FlatList } from 'react-native';
 import { Content, Container } from 'native-base';
 import { connect } from 'react-redux';
 import styles from './styles';
+import colors from '../../../utils/colors';
 import { getSeriesEpisodes, getEpisodesByPage } from '../../../api/home';
+import YouTube, { YouTubeStandaloneIOS, YouTubeStandaloneAndroid } from 'react-native-youtube';
 //customs
 import DetailComponent from '../../../components/DetailCard';
 import FooterScrollableTab from '../../../components/FooterScrollableTab';
 import GeneralDetails from '../../../components/GeneralDetails';
 import TechnicalDetails from '../../../components/TechnicalDetails';
 import EpisodesCard from '../../../components/EpisodesCard';
+import Loading from '../../../components/LoadingSpinner';
+import CustomModal from '../../../components/Modal';
 class Detail extends Component {
   state = {
     loading: false,
     cardInformation: this.props.navigation.getParam('cardInformation', []),
     episodes: [],
     urlNextPage: null,
-    isLoadingNext: false
+    isLoadingNext: false,
+    modalState: false,
+    titleState: '',
+    subtitleState: '',
+    isReady: false,
+    status: null,
+    error: null,
+    isPlaying: true,
+    isLooping: true,
   }
+  _youTubeRef = React.createRef();
   componentDidMount = async () => {
+    let { cardInformation } = this.state;
     try {
-      this.getEpisodesOfSeries();
+      if (cardInformation.type === 'anime') {
+        this.getEpisodesOfSeries();
+      }
     } catch (err) {
       console.log(err)
     }
@@ -34,7 +50,6 @@ class Detail extends Component {
       this.setState({
         episodes: getEpisodes.data.data,
         urlNextPage: getEpisodes.data.links.next,
-
       })
       console.log(getEpisodes, 'getEpisodes')
     } catch (err) {
@@ -55,9 +70,11 @@ class Detail extends Component {
         this.setState({
           episodes: nextArray,
           urlNextPage: response.data.links.next,
-          isLoadingNext: false,
         });
       }
+      this.setState({
+        isLoadingNext: false
+      });
     } catch (err) {
       console.log(err);
       this.setState({
@@ -74,18 +91,61 @@ class Detail extends Component {
     }
   };
 
+  onPlayVideo = (isReady) => {
+    let { cardInformation } = this.state;
+    this.setState({
+      isReady,
+    })
+  };
+
+  showModal = (modal, title, subtitle) => {
+    this.setState({
+      modalState: modal,
+      titleState: title,
+      subtitleState: subtitle,
+      isReady: true
+    });
+  };
+
+  onPressVideo = () => {
+    let { cardInformation } = this.state;
+    cardInformation.attributes.youtubeVideoId ?
+      this.showModal(false)
+      :
+      this.showModal(true, 'Sorry', 'No video available.')
+  }
+
   render() {
-    let { cardInformation, episodes } = this.state;
+    let { cardInformation, episodes, isLoadingNext, isReady, isPlaying, isLooping, modalState, titleState, subtitleState } = this.state;
     console.log(cardInformation, 'cardInformationdskfjsldk')
     return (
       <View style={styles.container}>
-        <StatusBar translucent backgroundColor="transparent" />
         <Container style={styles.backgroundContainer}>
           <Content showsVerticalScrollIndicator={false}>
             <DetailComponent
+              onPress={() => this.onPressVideo()}
+              onPLayVideo={cardInformation.attributes.youtubeVideoId ? isReady : false}
               name={cardInformation.attributes.canonicalTitle}
               picture={this.imageExists(cardInformation.attributes.posterImage)}
-              mainPicture={cardInformation.attributes.posterImage.large}
+              youtubeVideo={<YouTube
+                ref={this._youTubeRef}
+                apiKey={"AIzaSyBOJCZo6sFEM1u7zExOSXN6_BMAaqS0yXg"}
+                videoId={cardInformation.attributes.youtubeVideoId && cardInformation.attributes.youtubeVideoId}
+                play={isPlaying}
+                // loop={isLooping}
+                controls={1}
+                style={styles.youtubeVideoStyle}
+                onError={e => {
+                  console.log(e, 'youtube error')
+                }}
+                onReady={e => {
+                  this.setState({ isReady: true })
+                }}
+                onChangeState={e => {
+                  this.setState({ status: e.state });
+                }}
+              />}
+              mainPicture={this.imageExists(cardInformation.attributes.posterImage)}
               numberOfEpisodes={cardInformation.type === 'manga' ? cardInformation.attributes.chapterCount : cardInformation.attributes.episodeCount}
               minutesPerEpisode={cardInformation.type === 'manga' ? cardInformation.attributes.volumeCount : cardInformation.attributes.episodeLength}
               mangaTypeActive={cardInformation.type === 'manga' ? true : false}
@@ -106,30 +166,51 @@ class Detail extends Component {
                   status={cardInformation.attributes.status}
                 />}
             />
-            <View>
-              <Text style={styles.title}>Episodes</Text>
-              <FlatList
-                keyExtractor={i => i.id}
-                data={episodes}
-                onEndReached={this.loadMoreData}
-                onEndReachedThreshold={0.5}
-                // initialNumToRender={5}
-                maxToRenderPerBatch={3}
-                showsVerticalScrollIndicator={false}
-                renderItem={({ item, index }) => (
-                  <EpisodesCard
-                    picture={this.imageExists(item.attributes.thumbnail)}
-                    episodeName={item.attributes.canonicalTitle}
-                    season={item.attributes.seasonNumber}
-                    episode={item.attributes.number}
-                    // description={item.attributes.synopsis}
-                    airingDate={item.attributes.airdate}
-                  />
-                )
-                }
-              />
-            </View>
+            {cardInformation.type === 'anime' && (
+              <View>
+                <Text style={styles.title}>Episodes</Text>
+                <FlatList
+                  keyExtractor={i => i.id}
+                  data={episodes}
+                  onEndReached={this.loadMoreData}
+                  onEndReachedThreshold={0.5}
+                  // initialNumToRender={5}
+                  maxToRenderPerBatch={3}
+                  ListFooterComponent={
+                    <>
+                      {isLoadingNext && (
+                        <Loading color={'#8055E3'} />
+                      )}
+                    </>
+                  }
+                  showsVerticalScrollIndicator={false}
+                  renderItem={({ item, index }) => {
+                    if (item.attributes.canonicalTitle) {
+                      return <EpisodesCard
+                        picture={this.imageExists(item.attributes.thumbnail)}
+                        episodeName={item.attributes.canonicalTitle}
+                        season={item.attributes.seasonNumber}
+                        episode={item.attributes.number}
+                        // description={item.attributes.synopsis}
+                        airingDate={item.attributes.airdate}
+                      />;
+                    }
+                    else if (index === 0) {
+                      return <Text style={styles.title}>--</Text>
+                    }
+                  }
+                  }
+                />
+              </View>)}
           </Content>
+          <CustomModal
+            show={modalState}
+            title={titleState}
+            subtitle={subtitleState}
+            onPress={() => this.showModal(false)}
+            buttonText={'Accept'}
+            onTapOutside={() => this.showModal(false)}
+          />
         </Container>
       </View>
     );