+++ src/screens/app/Home/Category/styles.js
@@ -1,7 +1,7 @@
 import { StyleSheet } from 'react-native';
-import { resize } from '../../utils/styles';
-import colors from '../../utils/colors';
-import fonts from '../../utils/fonts';
+import { resize } from '../../../../utils/styles';
+import colors from '../../../../utils/colors';
+import fonts from '../../../../utils/fonts';
 export default StyleSheet.create({
   container: {
     flex: 1,