+++ src/screens/app/Home/SearchResult/styles.js
@@ -1,13 +1,13 @@
 import { StyleSheet, Platform } from 'react-native';
 import {
     resize,
-} from '../../utils/styles';
-import fonts from '../../utils/fonts';
-import colors from '../../utils/colors';
+} from '../../../../utils/styles';
+import fonts from '../../../../utils/fonts';
+import colors from '../../../../utils/colors';
 
 export default StyleSheet.create({
     container: {
-        alignItems: 'center',
+        // alignItems: 'center',
         backgroundColor: colors.backgroundColorApp,
     },
 });